+++ src/test/java/com/caoccao/javet/interop/converters/TestJavetProxyConverter.java
@@ -94,6 +94,9 @@ public class TestJavetProxyConverter extends BaseTestJavetRuntime {
         assertEquals("3", map.get("z"));
         assertEquals("4", v8Runtime.getExecutor("map.z = '4'; map.z;").executeString());
         assertEquals("4", map.get("z"));
+        assertEquals(
+                "[\"x\",\"y\",\"z\"]",
+                v8Runtime.getExecutor("JSON.stringify(Object.getOwnPropertyNames(map));").executeString());
         v8Runtime.getGlobalObject().delete("map");
         v8Runtime.lowMemoryNotification();
         javetProxyConverter.getConfig().setProxyMapEnabled(false);